deal with more update issues
don't update the message list, which forces us into view-recv or
view-sent mode, if we're not already in one of those modes.
also, sometimes there's a lag before the eventlogs update... so
if a message should exist, but we don't get a log entry for it,
indicate that the message data is unavailable.