made the build system not-so brain dead
now making does not modify index.html; and while commiting does
modify autoversion.js, that file is not under got. but more importantly,
although the modified autoversion.js propys a rebuild of index.js, and
that rebuild prompts a new set-timestamp, the new timestampt does not
affect the index.html that is under git. in the past it did affect
index.html, which would require a re-commit, which would invalidate the
autoversion.