audit issue rectification
item 2. added a "via" address to the MessageEvent. the via
address is set to the address of the trusted contract (or
zero in case the fromAddr is msg.sender). in this way a DApp
can indicate the via address to the recipient when the
message was not sent directly from the sender.